# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "用于异步抓取 Google Maps 数据的 API",
        "title": "Google Maps 抓取 Pro API",
        "contact": {
            "name": "API 支持",
            "url": "https://github.com/gosom/google-maps-scraper"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/health": {
            "get": {
                "description": "检查 API 服务是否健康",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "健康检查",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.HealthCheckResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/jobs": {
            "get": {
                "description": "分页列出所有抓取任务，支持按状态过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scrape"
                ],
                "summary": "获取任务列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "按状态过滤 (available, running, completed, cancelled, discarded, pending, retryable, scheduled)",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "返回的任务数量（默认：20，最大：100）",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用于分页的游标",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ListJobsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/jobs/{job_id}": {
            "get": {
                "description": "获取抓取任务的状态和结果",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scrape"
                ],
                "summary": "获取任务状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.JobStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "将任务及其结果放入删除队列。运行中的任务会被优先取消。",
                "tags": [
                    "scrape"
                ],
                "summary": "删除任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/api.DeleteJobResponse"
                        }
                    },
                    "400": {
                        "description": "无效的任务 ID",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "未找到指定任务",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/scrape": {
            "post": {
                "description": "提交一个新的 Google Maps 抓取任务进行处理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scrape"
                ],
                "summary": "提交抓取任务",
                "parameters": [
                    {
                        "description": "抓取请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ScrapeRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/api.ScrapeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.DeleteJobResponse": {
            "description": "成功将任务加入删除队列后返回的响应",
            "type": "object",
            "properties": {
                "message": {
                    "description": "状态信息",
                    "type": "string",
                    "example": "deletion queued"
                }
            }
        },
        "api.ErrorResponse": {
            "description": "请求失败时返回的错误响应",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        },
        "api.HealthCheckResponse": {
            "description": "指示服务状态的健康检查响应",
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "api.JobStatusResponse": {
            "description": "抓取任务的详细状态（完成后将包含结果）",
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "任务完成时间",
                    "type": "string",
                    "example": "2024-01-15T10:35:00Z"
                },
                "created_at": {
                    "description": "任务创建时间",
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "error": {
                    "description": "任务失败时的错误信息",
                    "type": "string",
                    "example": ""
                },
                "job_id": {
                    "description": "唯一任务标识符",
                    "type": "string",
                    "example": "kYzR8xLmNpQvWjX3"
                },
                "keyword": {
                    "description": "用于此任务的搜索关键词",
                    "type": "string",
                    "example": "restaurants in New York"
                },
                "result_count": {
                    "description": "找到的结果数量",
                    "type": "integer",
                    "example": 25
                },
                "results": {
                    "description": "抓取结果（地点数据数组）"
                },
                "started_at": {
                    "description": "任务开始处理的时间",
                    "type": "string",
                    "example": "2024-01-15T10:30:05Z"
                },
                "status": {
                    "description": "当前任务状态 (pending, running, completed, failed)",
                    "type": "string",
                    "example": "completed"
                }
            }
        },
        "api.JobSummary": {
            "description": "不包含完整结果数据的任务摘要",
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "任务完成时间",
                    "type": "string",
                    "example": "2024-01-15T10:35:00Z"
                },
                "created_at": {
                    "description": "任务创建时间",
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "error": {
                    "description": "任务失败时的错误信息",
                    "type": "string",
                    "example": ""
                },
                "job_id": {
                    "description": "唯一任务标识符",
                    "type": "string",
                    "example": "kYzR8xLmNpQvWjX3"
                },
                "keyword": {
                    "description": "用于此任务的搜索关键词",
                    "type": "string",
                    "example": "restaurants in New York"
                },
                "result_count": {
                    "description": "找到的结果数量",
                    "type": "integer",
                    "example": 25
                },
                "started_at": {
                    "description": "任务开始处理的时间",
                    "type": "string",
                    "example": "2024-01-15T10:30:05Z"
                },
                "status": {
                    "description": "当前任务状态",
                    "type": "string",
                    "example": "completed"
                }
            }
        },
        "api.ListJobsResponse": {
            "description": "带有下一页游标的分页任务列表",
            "type": "object",
            "properties": {
                "has_more": {
                    "description": "是否有更多结果",
                    "type": "boolean",
                    "example": true
                },
                "jobs": {
                    "description": "任务列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.JobSummary"
                    }
                },
                "next_cursor": {
                    "description": "下一页的游标（如果没有更多结果则为空）",
                    "type": "string",
                    "example": "eyJpZCI6MTIzfQ=="
                }
            }
        },
        "api.ScrapeRequest": {
            "description": "提交抓取任务的请求体",
            "type": "object",
            "properties": {
                "email": {
                    "description": "是否从网站中提取电子邮件地址",
                    "type": "boolean",
                    "example": false
                },
                "extra_reviews": {
                    "description": "提取更多评论",
                    "type": "boolean",
                    "example": false
                },
                "fast_mode": {
                    "description": "使用快速模式（使用隐身 HTTP 请求代替浏览器）",
                    "type": "boolean",
                    "example": false
                },
                "geo_coordinates": {
                    "description": "地理坐标，格式为 \"纬度,经度\"",
                    "type": "string",
                    "example": "40.7128,-74.0060"
                },
                "keyword": {
                    "description": "搜索关键词（例如：\"纽约的餐厅\"）",
                    "type": "string",
                    "example": "restaurants in New York"
                },
                "lang": {
                    "description": "结果语言代码（默认：\"zh-CN\"）",
                    "type": "string",
                    "example": "zh-CN"
                },
                "max_depth": {
                    "description": "分页的最大深度（默认：1，最大：100）",
                    "type": "integer",
                    "example": 1
                },
                "radius": {
                    "description": "搜索半径（公里）",
                    "type": "number",
                    "example": 5
                },
                "timeout": {
                    "description": "任务超时时间，单位秒（1-300，默认：300）",
                    "type": "integer",
                    "example": 300
                },
                "zoom": {
                    "description": "地图搜索缩放级别 (1-21)",
                    "type": "integer",
                    "example": 14
                }
            }
        },
        "api.ScrapeResponse": {
            "description": "成功提交抓取任务后返回的响应",
            "type": "object",
            "properties": {
                "job_id": {
                    "description": "唯一任务标识符",
                    "type": "string",
                    "example": "kYzR8xLmNpQvWjX3"
                },
                "status": {
                    "description": "当前任务状态",
                    "type": "string",
                    "example": "pending"
                }
            }
        }
    }
}